distro: allow user and group customizations for edge-installer
Followup from, f34380d5b5d2de0b8b75b7abfad095c255be5577 and
3a1765a5a89f912905bdf029d2f462bfd6b8c869, copied to the rest of the RHEL
distro definitions.

For now, these customizations have no effect on the manifest.

The new `with-users` variants of the edge-installer test cases include
the user customizations in the blueprint, but the manifests are
(currently) the same as the corresponding base cases.