osbuild-service-maintenance: Query job chains from root nodes
Instead of starting from osbuild and travelling up, start from root
nodes. If multiple osbuild jobs per depsolve are ever implemented, this
will not break.

Furthermore cascade deletes from jobs to job_dependencies. This is just
restricted to the whatever entries depend on, or are dependencies of,
the job to be deleted. So the amount of rows is small.