distro/rhel90beta: alias to rhel-90
Remove the rhel 90 beta code, which has fallen behind in terms of
features and bug fixes and make it an alias to 9.0, i.e. it can
still be requested but it will end up using the rhel 9.0 image
definitions. Also remove the manifests and the entries in the
manifest generator.