test: disable full image building in generation.sh
The purpose of the test is to check that the dnf-json socket can be
started automatically when running the test case generator while the
service or socket isn't enabled/started.

dnf-json will still be used to depsolve the packages and create the
manifest even if the image is not built.