composer: fail to start if not listening on any API socket.
osbuild-composer can now start and keep running without listening on any
API socket. This effectively means that there is no way to communicate
with such instance of osbuild-composer.

Change osbuild-composer to exit with error in case no API socket has been
activated. The log message is intended for the on-premise use case, since
this is the situation when users tend to start osbuild-composer.service
manually and are observing a weird behavior.

Related to https://github.com/osbuild/osbuild-composer/pull/2550