osbuild1: remove UnmarshalJSON for Stage and StageOptions
OSBuild used to return the stage options as part of the result object
for v1 manifests.  We didn't use this information anywhere.  Currently
we convert v1 results to the v2 format while parsing the results of jobs
from old manifests (old distro definitions), but the StageOptions are
ignored and we only care about the StageMetadata.